% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_jud.R
\name{aux_jud}
\alias{aux_jud}
\title{Sub-índices para cálculo do índice Judiciário.}
\usage{
aux_jud(df, group_name = "coletiva", min_dt, max_dt)
}
\arguments{
\item{df}{Dataframe contendo as colunas \code{date}, \code{gov}, \code{index}, \code{category}.}

\item{group_name}{String contendo o grupo que será utilizado para calcular o sub-índice.}
}
\value{

}
\description{
Função para calcular sub-índices necessários para o cálculo do índice Judiciário.
}
\details{
O arquivo de input deve corresponder a base de dados já limpa, contendo colunas indicando ocorrência das
Ações Diretas de Inconstitucionalidades (ADIs) julgadas pelo poder Judiciário.
Cada linha representa uma ação julgada que foi solicitada e finalizada dentro de um mesmo governo:
O arquivo deve conter as seguintes colunas:
\code{date} - coluna de datas no formato YYYY/MM/DD;
\code{gov} - Governo no qual houve o julgamento da ADIs;
\code{index} - Decisão tomada pelo Judiciário (contra ou favorável);
\code{category} - Organização do Judiciário que tomou a decisão (coletivo ou individual);
}
\examples{
\dontrun{
calc_sub_indice_judiciario(df = df_cleaned, group_name = 'individual')
}

}
\author{
Gabriel Bellé & Luiz Paulo
}
