% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_jud_clean.R
\name{aux_jud_clean}
\alias{aux_jud_clean}
\title{Auxiliar na limpeza base de dados judiciário}
\usage{
aux_jud_clean(db = as.data.frame(), depara = as.data.frame())
}
\description{
Auxiliar na limpeza base de dados judiciário
}
\author{
Luiz Paulo Tavares
}
