% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igov_jud.R
\name{igov_jud}
\alias{igov_jud}
\title{Índice Judiciário.}
\usage{
igov_jud(df)
}
\arguments{
\item{df}{Dataframe contendo as colunas \code{date}, \code{gov}, \code{index}, \code{category}.}
}
\value{

}
\description{
Função para o cálculo do índice Judiciário.
}
\details{
O arquivo de input deve corresponder a base de dados já limpa, contendo colunas indicando ocorrência das
Ações Diretas de Inconstitucionalidades (ADIs) julgadas pelo poder Judiciário.
Cada linha representa uma ação julgada que foi solicitada e finalizada dentro de um mesmo governo:
O arquivo deve conter as seguintes colunas:
\code{date} - Coluna de datas no formato YYYY/MM/DD;
\code{gov} - Governo no qual houve o julgamento da ADIs;
\code{index} - Decisão tomada pelo Judiciário (contra ou favor);
\code{category} - Organização do Judiciário que tomou a decisão (coletiva ou individual);
}
\examples{
\dontrun{

calc_judiciario(df = df_cleaned, peso = 2)

}

}
\author{
Gabriel Bellé & Luiz Paulo
}
