% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igov_legis.R
\name{igov_legis}
\alias{igov_legis}
\title{Índice Legislativo}
\usage{
igov_legis(df, peso)
}
\arguments{
\item{df}{Dataframe contendo as colunas \code{date}, \code{analisada}, \code{pura}, \code{emendada}, \code{sem_eficacia}, \code{gov}.}

\item{peso}{Constante indicando a relação da importância entre MPs puras e MPs emendadas para o índice.}
}
\value{
O retorno é um dataframe agrupado por dia e governo contendo as colunas:

\code{date};

\code{gov};

\code{soma_movel_analisada};

\code{soma_movel_pura};

\code{soma_movel_emendada};

\code{soma_movel_sem_eficacia};

\code{value_analisada};

\code{value_pura};

\code{value_emendada};

\code{value_sem_eficacia};

\code{pct_pura};

\code{pct_emendada};

\code{pct_sem_eficacia};

\code{pura_peso};

\code{emendada_peso};

\code{indice_legislativo}.
}
\description{
Função para calcular o índice Legislativo.
}
\details{
O arquivo de input deve corresponder a base de dados já limpa, contendo colunas indicando a quantidade de
Medidas Provisórias (MPs) julgadas pelo poder Legislativo, classificadas em:

\code{date} - coluna de datas no formato YYYY/MM/DD;

\code{gov} - Governo no qual houve o julgamento da MPs;

\code{analisada} - Quantidade total de MPs analisadas;

\code{pura} - Quantidade de MPs aprovadas sem modificação;

\code{emendada} - Quantidade de MPs aprovadas com modificação;

\code{sem_eficacia} - Quantidade de MPs não aprovadas.
}
\examples{

\dontrun{

Igov4i::igov_legis(df = dataset_legis, peso = 2)

}

}
\author{
Gabriel Bellé & Luiz Paulo Tavares
}
