% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igov_pop.R
\name{igov_pop}
\alias{igov_pop}
\title{Índice de Popularidade}
\usage{
igov_pop(df = as.data.frame())
}
\arguments{
\item{df}{Dataframe contendo as colunas \code{date}, \code{instituto}, \code{presidente} e o grau de aprovação da sociedade civil:

\code{ótimo/bom};

\code{regular};

\code{ruim/péssimo};

\code{não sabe}.}
}
\value{
O retorno é um dataframe agrupado por mês e governo contendo as colunas:

\code{date};

\code{media_bom};

\code{media_regular};

\code{indice_pop}.
}
\description{
Função para calcular o índice de Opinião Pública, isto é, mensurar o desempenho do governo com relação a sociedade civil.
}
\details{
O arquivo de input deve corresponder a base de dados já limpa, contendo colunas indicando as pesquisas de opinião referente aos institutos de pesquisa.
}
\examples{

\dontrun{

igov_pop <- Igov4i::igov_pop(df = pesq_pop)

}

}
\author{
Gabriel Bellé & Luiz Paulo Tavares
}
