% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jud_clean.R
\name{jud_clean}
\alias{jud_clean}
\title{Limpeza base de dados judiciário}
\usage{
jud_clean(db = as.data.frame(), depara = as.data.frame())
}
\description{
Limpeza base de dados judiciário
}
\author{
Luiz Paulo Tavares
}
